FEAT: Implement susbcribe_filter to prefilter pending transactions
